Rename dpz to pdz and unpack less aggressively.
Fix #45 and fix #52.
This fixes two bugs. The first was that PdParty for iOS expects .pdz not
.dpz as an extension, so this change makes PdDroidParty match PdParty
behaviour. The second is that PdDroidParty was prompting to open
absolutely every type of file ever.
See note on #52 for why this now means you cannot email pdz files around
(only download in the browser).